split internal
